修复一些问题

+++ docker/nginx/docker-entrypoint.sh
@@ -3,10 +3,10 @@
 if [ "${NGINX_HTTPS_ENABLED}" = true ]; then
     # Check if the certificate and key files for the specified domain exist
     if [ -n "${CERTBOT_DOMAIN}" ] && \
-       [ -f "/etc/letsencrypt/live/${CERTBOT_DOMAIN}/${NGINX_SSL_CERT_FILENAME}" ] && \
-       [ -f "/etc/letsencrypt/live/${CERTBOT_DOMAIN}/${NGINX_SSL_CERT_KEY_FILENAME}" ]; then
-        SSL_CERTIFICATE_PATH="/etc/letsencrypt/live/${CERTBOT_DOMAIN}/${NGINX_SSL_CERT_FILENAME}"
-        SSL_CERTIFICATE_KEY_PATH="/etc/letsencrypt/live/${CERTBOT_DOMAIN}/${NGINX_SSL_CERT_KEY_FILENAME}"
+       [ -f "/etc/letsencrypt/live/${CERTBOT_DOMAIN}/fullchain.pem" ] && \
+       [ -f "/etc/letsencrypt/live/${CERTBOT_DOMAIN}/privkey.pem" ]; then
+        SSL_CERTIFICATE_PATH="/etc/letsencrypt/live/${CERTBOT_DOMAIN}/fullchain.pem"
+        SSL_CERTIFICATE_KEY_PATH="/etc/letsencrypt/live/${CERTBOT_DOMAIN}/privkey.pem"
     else
         SSL_CERTIFICATE_PATH="/etc/ssl/${NGINX_SSL_CERT_FILENAME}"
         SSL_CERTIFICATE_KEY_PATH="/etc/ssl/${NGINX_SSL_CERT_KEY_FILENAME}"
@@ -14,11 +14,14 @@ if [ "${NGINX_HTTPS_ENABLED}" = true ]; then
     export SSL_CERTIFICATE_PATH
     export SSL_CERTIFICATE_KEY_PATH
 
-    # set the HTTPS_CONFIG environment variable to the content of the https.conf.template
-    HTTPS_CONFIG=$(envsubst < /etc/nginx/https.conf.template)
-    export HTTPS_CONFIG
     # Substitute the HTTPS_CONFIG in the default.conf.template with content from https.conf.template
-    envsubst '${HTTPS_CONFIG}' < /etc/nginx/conf.d/default.conf.template > /etc/nginx/conf.d/default.conf
+    if [ -f "${SSL_CERTIFICATE_PATH}" ] && [ -f "${SSL_CERTIFICATE_KEY_PATH}" ]; then
+        "${SSL_HTTP2}" == true && SSL_HTTP2="http2 on;" || SSL_HTTP2=""
+        # set the HTTPS_CONFIG environment variable to the content of the https.conf.template
+        HTTPS_CONFIG=$(envsubst < /etc/nginx/https.conf.template)
+        export HTTPS_CONFIG
+        envsubst '${HTTPS_CONFIG}' < /etc/nginx/conf.d/default.conf.template > /etc/nginx/conf.d/default.conf
+    fi
 fi
 
 if [ "${NGINX_ENABLE_CERTBOT_CHALLENGE}" = "true" ]; then
